[FAST PUSH] Goodrich Service (#157)
This is a fast push that will skip over any review process in order to get the MVP to market! May contain bugs that will need to be hotfixed.

WSO+Goodrich frontend implementation for May 2021. This code represents the MVP and should be treated as such: more features will come over time. All this code was written in a span of 3 days, and that is reflected. Please email al15@williams.edu or aidanlloydtucker@gmail.com for questions.