fix: Fix various bugs
Fix a bug where refresh leads to login page.
It is because of a delay between loading of API token and user.
I have temporarily worked around it, but better redux logic is needed in the future.

In DormtrakFacts, `graph_id` is commented out because it somehow leads to the chart being smaller.

Some navigation errors in DormTrakRecentComments are also fixed.